GitHubProjects_RRR_V5/road-freight-risk-ai-v3-main/requirements.txt
streamlit==1.35.0
streamlit-folium==0.16.0
streamlit-js-eval==0.1.5
folium==0.14.0
branca==0.6.0
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.1

# Compatibility notes:
# - folium 0.14.0 is stable and widely compatible
# - streamlit-folium 0.16.0 works well with folium 0.14.0
# - branca 0.6.0 is required for folium 0.14.0 color mapping
